Extracted load_gnps functions from metabolomics
